chore: utils for injected script
